// src/types/generated/FileEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FileKind } from "./FileKind";

export type FileEntry = { name: string, path: string, kind: FileKind, 
/**
 * For symlinks: whether the target is a directory.
 */
linkIsDir: boolean, size: number, modified: number | null, permissions: number | null, owner: string | null, group: string | null, hidden: boolean, };